Fix: use pointers in profileContainer for intent(in).
Change from using allocatables to pointers for profileEvents and nProfileEvents in fclProfileContainer.
This is needed since the correct intent for buffer on overloaded assignment(=) when reading from device is intent(in) which means the allocatable components cannot be modified. Therefore pointers are the correct way to go here.
FIxes a failing test in intel fortran which picked this up and should fix a semantic parse error in f18.